Initial GPU vector rendering as a sample (#524)

+++ kmem/src/commonMain/kotlin/com/soywiz/kmem/Numbers.kt
@@ -25,9 +25,9 @@ public fun Float.toIntFloor(): Int = floor(this).toInt()
 public fun Double.toIntFloor(): Int = floor(this).toInt()
 
 /** Converts [this] into [Int] rounding to the nearest */
-public fun Float.toIntRound(): Int = round(this).toLong().toInt()
+public fun Float.toIntRound(): Int = round(this).toInt()
 /** Converts [this] into [Int] rounding to the nearest */
-public fun Double.toIntRound(): Int = round(this).toLong().toInt()
+public fun Double.toIntRound(): Int = round(this).toInt()
 
 /** Convert this [Long] into an [Int] but throws an [IllegalArgumentException] in the case that operation would produce an overflow */
 public fun Long.toIntSafe(): Int = if (this in Int.MIN_VALUE.toLong()..Int.MAX_VALUE.toLong()) this.toInt() else throw IllegalArgumentException("Long doesn't fit Integer")
